fix naming to reflect that this is now a transform not a formatter
